End of GSOC release Fix converting projects with gtk-sharp reference but without gtk-gui folder * Makefile.am: * libsteticui/Application.cs: * MonoDevelop.GtkCore2.csproj: * MonoDevelop.GtkCore.addin.xml: * libsteticui/ProjectBackend.cs: * MonoDevelop.GtkCore/GtkDesignInfo.cs: * MonoDevelop.GtkCore.NodeBuilders/WindowsFolder.cs: * MonoDevelop.GtkCore.NodeBuilders/WidgetNodeBuilder.cs: * MonoDevelop.GtkCore.NodeBuilders/WindowsFolderNodeBuilder.cs: * MonoDevelop.GtkCore.NodeBuilders/ProjectFileNodeBuilderExtension.cs:

+++ stetic2/MonoDevelop.GtkCore2/MonoDevelop.GtkCore/GtkDesignInfo.cs
@@ -172,7 +172,7 @@ namespace MonoDevelop.GtkCore
 		}
 		
 		public bool OldVersion {
-			get { return project.IsFileInProject (SteticFile); }
+			get { return project.IsFileInProject (SteticFile) || wasConverted == false; }
 		}
 		
 		FilePath ObjectsFile {
@@ -350,7 +350,8 @@ namespace MonoDevelop.GtkCore
 			if (File.Exists (oldGeneratedFile))
 				FileService.DeleteFile (oldGeneratedFile);
 				
-			FileService.DeleteDirectory (oldGuiFolder);
+			if (Directory.Exists (oldGuiFolder))
+				FileService.DeleteDirectory (oldGuiFolder);
 		}
 		
 		public bool UpdateGtkFolder ()
